requirements.txt
fastapi
starlette
uvicorn
python-dotenv
supabase
requests
pydantic
typing_extensions
fastmcp
